tried semaphores but analog values random.
