+++ platformio.ini
@@ -88,6 +88,6 @@ lib_deps =
 
 ;-------------------------------------------------------
 [dir]
-main = +<main/>
-test_comms = +<test/test_comms/>
-test_comms_and_display = +<test/comms_and_display/>
+main = +<main/> -<test_comms/> -<comms_and_display/>
+test_comms = +<test_comms/>
+test_comms_and_display = +<comms_and_display/>
